Avalon

+++ Avalon/Avalon.Draw/Color.cs
@@ -2,8 +2,8 @@ namespace Avalon.Draw;
 
 public class Color : Any
 {
-    public virtual int Alpha { get; set; }
-    public virtual int Red { get; set; }
-    public virtual int Green { get; set; }
-    public virtual int Blue { get; set; }
+    public virtual long Alpha { get; set; }
+    public virtual long Red { get; set; }
+    public virtual long Green { get; set; }
+    public virtual long Blue { get; set; }
 }
